tests moved to strings/setintegrationTest
